refactor: extract ios manager tests
