+++ src/managers/ios/mod.rs
@@ -486,321 +486,4 @@ impl crate::managers::common::UnifiedDeviceManager for IosManager {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    #[cfg(target_os = "macos")]
-    fn test_extract_ios_version() {
-        // Test version parse logic (actual implementation: major + minor/100 + patch/10000)
-        assert_eq!(extract_ios_version("iOS 18.5"), 18.05); // 18 + 5/100
-        assert_eq!(extract_ios_version("iOS 17.0"), 17.0); // 17 + 0/100
-        assert_eq!(extract_ios_version("iOS 16.4"), 16.04); // 16 + 4/100
-
-        // With patch version: major + minor/100 + patch/10000
-        assert!((extract_ios_version("iOS 15.2.1") - 15.0201).abs() < 0.0001); // Consider floating point precision
-        assert!((extract_ios_version("iOS 16.3.2") - 16.0302).abs() < 0.0001);
-
-        // Other OS follow the same logic
-        assert_eq!(extract_ios_version("watchOS 10.0"), 10.0);
-        assert_eq!(extract_ios_version("tvOS 17.2"), 17.02);
-
-        // Edge cases
-        assert_eq!(extract_ios_version("iOS-17-0"), 17.0);
-        assert_eq!(extract_ios_version("iOS"), 0.0);
-        assert_eq!(extract_ios_version(""), 0.0);
-
-        // Verify sort functionality
-        assert!(extract_ios_version("iOS 18.5") > extract_ios_version("iOS 18.0"));
-        assert!(extract_ios_version("iOS 17.9") > extract_ios_version("iOS 17.1"));
-    }
-
-    #[test]
-    #[cfg(target_os = "macos")]
-    fn test_parse_device_type_display_name() {
-        // Test basic device type conversion (not dependent on implementation details)
-        let result = IosManager::parse_device_type_display_name(
-            "com.apple.CoreSimulator.SimDeviceType.iPhone-15-Pro",
-        );
-        assert!(result.contains("iPhone"));
-        assert!(result.contains("15"));
-        assert!(result.contains("Pro"));
-
-        let result2 = IosManager::parse_device_type_display_name("iPhone-14");
-        assert!(result2.contains("iPhone"));
-        assert!(result2.contains("14"));
-
-        // Test empty string
-        assert_eq!(IosManager::parse_device_type_display_name(""), "");
-    }
-
-    /// Test IosManager creation and initialization
-    #[test]
-    #[cfg(target_os = "macos")]
-    fn test_ios_manager_new() {
-        let _manager = IosManager::new().expect("Failed to create IosManager");
-        // Manager should be created successfully
-        // Note: IosManager doesn't implement Debug
-        // assert!(format!("{manager:?}").contains("IosManager"));
-    }
-
-    /// Test iOS device parsing edge cases
-    #[test]
-    #[cfg(target_os = "macos")]
-    fn test_ios_device_parsing_edge_cases() {
-        // Test version extraction with various formats
-        assert_eq!(extract_ios_version("iOS 18"), 18.0);
-        assert_eq!(extract_ios_version("iOS-18"), 18.0);
-        assert_eq!(extract_ios_version("iOS.18"), 18.0);
-        assert_eq!(extract_ios_version("invalid"), 0.0);
-        assert_eq!(extract_ios_version("iOS 99.99.99"), 99.9999);
-
-        // Test device type parsing with complex names
-        let result = IosManager::parse_device_type_display_name(
-            "com.apple.CoreSimulator.SimDeviceType.iPad-Pro-12-9-inch-6th-generation",
-        );
-        assert!(result.contains("iPad"));
-        assert!(result.contains("Pro"));
-
-        let result2 = IosManager::parse_device_type_display_name("Apple-Watch-Series-8-45mm");
-        assert!(result2.contains("Apple"));
-        assert!(result2.contains("Watch"));
-    }
-
-    /// Test iOS manager creation (simplified)
-    #[test]
-    fn test_ios_manager_creation() {
-        let _manager = IosManager::new().expect("Failed to create IosManager");
-        // Manager created successfully
-    }
-
-    /// Test iOS manager device operations (non-macOS behavior) - DISABLED
-    #[allow(dead_code)]
-    #[cfg(not(target_os = "macos"))]
-    async fn test_ios_manager_non_macos_operations_disabled() {
-        use crate::managers::common::{DeviceConfig, DeviceManager};
-        use std::collections::HashMap;
-
-        let _manager = IosManager::new().expect("Failed to create IosManager");
-
-        // All operations should fail on non-macOS
-        assert!(<IosManager as DeviceManager>::list_devices(&_manager)
-            .await
-            .is_err());
-        assert!(
-            <IosManager as DeviceManager>::start_device(&_manager, "test")
-                .await
-                .is_err()
-        );
-        assert!(
-            <IosManager as DeviceManager>::stop_device(&_manager, "test")
-                .await
-                .is_err()
-        );
-
-        let config = DeviceConfig {
-            name: "Test Device".to_string(),
-            device_type: "iPhone-15".to_string(),
-            version: "iOS 17.0".to_string(),
-            ram_size: None,
-            storage_size: None,
-            additional_options: HashMap::new(),
-        };
-        assert!(
-            <IosManager as DeviceManager>::create_device(&_manager, &config)
-                .await
-                .is_err()
-        );
-        assert!(
-            <IosManager as DeviceManager>::delete_device(&_manager, "test")
-                .await
-                .is_err()
-        );
-        assert!(
-            <IosManager as DeviceManager>::wipe_device(&_manager, "test")
-                .await
-                .is_err()
-        );
-    }
-
-    /// Test iOS manager UnifiedDeviceManager trait (non-macOS) - DISABLED
-    #[allow(dead_code)]
-    #[cfg(not(target_os = "macos"))]
-    async fn test_ios_manager_unified_device_manager_non_macos_disabled() {
-        use crate::managers::common::{DeviceConfig, UnifiedDeviceManager};
-        use std::collections::HashMap;
-
-        let _manager = IosManager::new().expect("Failed to create IosManager");
-
-        // All UnifiedDeviceManager operations should fail on non-macOS
-        assert!(
-            <IosManager as UnifiedDeviceManager>::list_devices(&_manager)
-                .await
-                .is_err()
-        );
-        assert!(
-            <IosManager as UnifiedDeviceManager>::start_device(&_manager, "test")
-                .await
-                .is_err()
-        );
-        assert!(
-            <IosManager as UnifiedDeviceManager>::stop_device(&_manager, "test")
-                .await
-                .is_err()
-        );
-
-        let config = DeviceConfig {
-            name: "Test Device".to_string(),
-            device_type: "iPhone-15".to_string(),
-            version: "iOS 17.0".to_string(),
-            ram_size: None,
-            storage_size: None,
-            additional_options: HashMap::new(),
-        };
-        assert!(
-            <IosManager as UnifiedDeviceManager>::create_device(&_manager, &config)
-                .await
-                .is_err()
-        );
-        assert!(
-            <IosManager as UnifiedDeviceManager>::delete_device(&_manager, "test")
-                .await
-                .is_err()
-        );
-        assert!(
-            <IosManager as UnifiedDeviceManager>::wipe_device(&_manager, "test")
-                .await
-                .is_err()
-        );
-        assert!(!<IosManager as UnifiedDeviceManager>::is_available(&_manager).await);
-    }
-
-    /// Test iOS device priority calculation - DISABLED (DeviceInfo::new not available)
-    #[allow(dead_code)]
-    #[cfg(target_os = "macos")]
-    fn test_ios_device_priority_disabled() {
-        // DeviceInfo::new method doesn't exist
-        // Test disabled until DeviceInfo API is available
-    }
-
-    /// Test iOS device status parsing
-    #[test]
-    #[cfg(target_os = "macos")]
-    fn test_ios_device_status_parsing() {
-        // Test with various status strings that might be returned by simctl
-
-        // These would be used in actual status parsing
-        let statuses = vec!["Booted", "Shutdown", "Creating", "Booting", "Shutting Down"];
-
-        for status in statuses {
-            // Status parsing would happen in the actual implementation
-            match status {
-                "Booted" => assert_eq!(status, "Booted"),
-                "Shutdown" => assert_eq!(status, "Shutdown"),
-                _ => assert!(!status.is_empty()),
-            }
-        }
-    }
-
-    /// Test iOS manager error handling - DISABLED
-    #[allow(dead_code)]
-    async fn test_ios_manager_error_handling_disabled() {
-        let _manager = IosManager::new().expect("Failed to create IosManager");
-
-        // Test that operations don't panic and return appropriate errors
-        #[cfg(not(target_os = "macos"))]
-        {
-            use crate::managers::common::DeviceConfig;
-            use std::collections::HashMap;
-
-            let config = DeviceConfig {
-                name: "Test".to_string(),
-                device_type: "iPhone".to_string(),
-                version: "17.0".to_string(),
-                ram_size: None,
-                storage_size: None,
-                additional_options: HashMap::new(),
-            };
-
-            // All operations should return errors, not panic
-            let result = _manager.create_device(&config).await;
-            assert!(result.is_err());
-            assert!(result.unwrap_err().to_string().contains("macOS"));
-        }
-    }
-
-    /// Test iOS version comparison functionality
-    #[test]
-    #[cfg(target_os = "macos")]
-    fn test_ios_version_comparison() {
-        // Test that version extraction supports proper comparison
-        let v1 = extract_ios_version("iOS 17.0");
-        let v2 = extract_ios_version("iOS 17.1");
-        let v3 = extract_ios_version("iOS 18.0");
-
-        assert!(v1 < v2);
-        assert!(v2 < v3);
-        assert!(v1 < v3);
-
-        // Test with patch versions
-        let v4 = extract_ios_version("iOS 17.0.1");
-        let v5 = extract_ios_version("iOS 17.0.2");
-
-        assert!(v4 < v5);
-        assert!(v1 < v4); // 17.0 < 17.0.1
-    }
-
-    /// Test device type display name formatting
-    #[test]
-    #[cfg(target_os = "macos")]
-    fn test_device_type_display_formatting() {
-        // Test that the function can parse device type names and format them appropriately
-        let test_cases = vec![
-            "iPhone-15-Pro-Max",
-            "iPad-Pro-11-inch",
-            "Apple-Watch-Series-9",
-            "Apple-TV-4K",
-        ];
-
-        for input in test_cases {
-            let result = IosManager::parse_device_type_display_name(input);
-            // Basic checks - result should not be empty and should contain some expected patterns
-            assert!(
-                !result.is_empty(),
-                "Result should not be empty for input '{input}'"
-            );
-
-            // Should contain some parts of the original input (transformed)
-            if input.contains("iPhone") {
-                assert!(
-                    result.contains("iPhone"),
-                    "Result '{result}' should contain 'iPhone' for input '{input}'"
-                );
-            }
-            if input.contains("iPad") {
-                assert!(
-                    result.contains("iPad"),
-                    "Result '{result}' should contain 'iPad' for input '{input}'"
-                );
-            }
-            if input.contains("Apple-Watch") {
-                assert!(
-                    result.contains("Apple"),
-                    "Result '{result}' should contain 'Apple' for input '{input}'"
-                );
-            }
-            if input.contains("TV") {
-                assert!(
-                    result.contains("TV"),
-                    "Result '{result}' should contain 'TV' for input '{input}'"
-                );
-            }
-
-            // The result should have spaces instead of hyphens for readability
-            assert!(
-                !result.contains("-"),
-                "Result '{result}' should not contain hyphens"
-            );
-        }
-    }
-}
+mod tests;
